Add generatorToCallback function to lib/function module.
It receives a generator function and returns callback function that
receives express request / response object and next callback.

If the result of generator function successfully resolved, that result is
directly sent by calling express res.send() method.

If there's an error, it catches the error and call next function as its
first argument.